Rename folder from push-notifications to notifications
