+++ src/notifications/index.ts
@@ -3,3 +3,4 @@ export {
   usePushNotificationsEnabled,
 } from './use-push-notifications-enabled';
 export {usePushNotifications} from './use-push-notifications';
+export {isConfigEnabled} from './utils';
