+++ src/notifications/use-register.tsx
@@ -1,4 +1,4 @@
-import {registerForPushNotifications} from '@atb/push-notifications/api';
+import {registerForPushNotifications} from './api';
 import {useMutation} from '@tanstack/react-query';
 
 export const useRegister = () => {
