+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_ProfileStack/Profile_DebugInfoScreen.tsx
@@ -54,7 +54,7 @@ import {useParkingViolationsReportingEnabledDebugOverride} from '@atb/parking-vi
 import {shareTravelHabitsSessionCountKey} from '@atb/beacons/use-maybe-show-share-travel-habits-screen';
 import {hasSeenShareTravelHabitsScreenKey} from '@atb/beacons/use-has-seen-share-travel-habits-screen';
 import {useAnnouncementsState} from '@atb/announcements';
-import {usePushNotificationsEnabledDebugOverride} from '@atb/push-notifications';
+import {usePushNotificationsEnabledDebugOverride} from '@atb/notifications';
 
 function setClipboard(content: string) {
   Clipboard.setString(content);
